requirements.txt
# Video Generation Pipeline Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# ElevenLabs TTS
elevenlabs>=0.2.28

# FAL.ai Image Generation
fal-client>=0.3.0

# YouTube API
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.108.0

# Video/Audio Processing
moviepy>=1.0.3
pydub>=0.25.1
ffmpeg-python>=0.2.1

# Data handling
pandas>=2.1.0
pyyaml>=6.0

# Logging and utilities
tqdm>=4.66.0
click>=8.1.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.10.0
flake8>=6.1.0
mypy>=1.5.0
